refactor: move useFilterQueryParams
