+++ src/lib/sectionList/filters/index.ts
@@ -1,2 +1,3 @@
 export * from './filterConfig'
 export * from './useSectionListFilters'
+export * from './useFilterQueryParams'
