+++ src/lib/sectionList/filters/useFilterQueryParams.ts
@@ -1,0 +1,13 @@
+import { useMemo } from 'react'
+import { useModelSectionHandleOrThrow } from '../../routeUtils'
+import { parseFiltersToQueryParams } from './parseFiltersToQueryParams'
+import { useSectionListFilters } from './useSectionListFilters'
+
+export const useFilterQueryParams = (): string[] => {
+    const [filters] = useSectionListFilters()
+    const section = useModelSectionHandleOrThrow()
+
+    return useMemo(() => {
+        return parseFiltersToQueryParams(filters, section)
+    }, [filters, section])
+}
