+++ src/lib/sectionList/useParamsForDataQuery.ts
@@ -1,17 +1,6 @@
 import { useMemo } from 'react'
-import { useModelSectionHandleOrThrow } from '../routeUtils'
-import { useSectionListFilters } from './filters'
-import { parseFiltersToQueryParams } from './filters/parseFiltersToQueryParams'
+import { useFilterQueryParams } from './filters'
 import { usePaginationQueryParams } from './usePaginationParams'
-
-export const useFilterQueryParams = (): string[] => {
-    const [filters] = useSectionListFilters()
-    const section = useModelSectionHandleOrThrow()
-
-    return useMemo(() => {
-        return parseFiltersToQueryParams(filters, section)
-    }, [filters, section])
-}
 
 export const useParamsForDataQuery = () => {
     const [paginationParams] = usePaginationQueryParams()
